Updated to geonames 1.0.1.
git-svn-id: https://wush.net/svn/mindalliance@1502 baad322d-9929-0410-88f0-f92e8ff3e1bd

+++ channels/trunk/src/main/java/com/mindalliance/channels/geo/GeoLocation.java
@@ -28,7 +28,7 @@ public class GeoLocation implements Serializable {
     private String stateCode;
     private String countyCode;
     private String cityCode;
-    private Integer population;
+    private long population;
     private double latitude;
     private double longitude;
     private String streetAddress;
@@ -105,7 +105,7 @@ public class GeoLocation implements Serializable {
         }
     }
 
-    private Integer getPopulation( Toponym topo ) {
+    private Long getPopulation( Toponym topo ) {
         try {
             return topo.getPopulation();
         } catch ( InsufficientStyleException e ) {
@@ -177,7 +177,7 @@ public class GeoLocation implements Serializable {
         this.cityCode = cityCode;
     }
 
-    public Integer getPopulation() {
+    public Long getPopulation() {
         return population;
     }
